TranslationTool/requirements.txt
pandas
google-genai
python-dotenv
tqdm
